# Fix omnisci array methods tests