Fix https://github.com/xnd-project/rbc/issues/79

Edit: Add information about runtime speedup

Without this PR:
```
$ pytest -sv --durations=0 rbc/tests/test_omnisci_array_methods.py -x --disable-warnings
========================================================================================================== test session starts ==========================================================================================================
platform linux -- Python 3.8.1, pytest-5.3.4, py-1.8.1, pluggy-0.13.1 -- /home/guilhermel/miniconda3/envs/rbc-dev/bin/python
cachedir: .pytest_cache
rootdir: /work/guilhermel/git/rbc
collecting ...  OmnisciDB version (5, 3, 0, 'dev-20200516-2d56e3d31b')
collected 14 items

rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[fill] PASSED
rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[max] PASSED
rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[max_empty] PASSED
rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[max_initial] PASSED
rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[mean] PASSED
rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[mean_empty_float] PASSED
rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[mean_empty_int] PASSED
rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[min] PASSED
rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[min_empty] PASSED
rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[min_initial] PASSED
rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[sum] PASSED
rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[sum_initial] PASSED
rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[prod] PASSED
rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[prod_initial] PASSED

======================================================================================================== slowest test durations =========================================================================================================
2.27s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[fill]
1.76s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[prod_initial]
1.73s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[max_initial]
1.71s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[sum_initial]
1.70s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[max]
1.69s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[min]
1.65s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[sum]
1.64s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[min_initial]
1.63s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[mean_empty_float]
1.61s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[min_empty]
1.60s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[mean_empty_int]
1.59s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[max_empty]
1.59s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[mean]
1.53s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[prod]
0.03s setup    rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[fill]
```

With the changes introduced in this PR:

```
$ pytest -sv --durations=0 rbc/tests/test_omnisci_array_methods.py -x --disable-warnings
========================================================================================================== test session starts ==========================================================================================================
platform linux -- Python 3.8.1, pytest-5.3.4, py-1.8.1, pluggy-0.13.1 -- /home/guilhermel/miniconda3/envs/rbc-dev/bin/python
cachedir: .pytest_cache
rootdir: /work/guilhermel/git/rbc
collecting ...  OmnisciDB version (5, 3, 0, 'dev-20200516-2d56e3d31b')
collected 14 items

rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[fill] PASSED
rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[max] PASSED
rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[max_empty] PASSED
rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[max_initial] PASSED
rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[mean] PASSED
rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[mean_empty_float] PASSED
rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[mean_empty_int] PASSED
rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[min] PASSED
rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[min_empty] PASSED
rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[min_initial] PASSED
rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[sum] PASSED
rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[sum_initial] PASSED
rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[prod] PASSED
rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[prod_initial] PASSED

======================================================================================================== slowest test durations =========================================================================================================
0.63s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[fill]
0.39s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[min_empty]
0.38s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[max_initial]
0.37s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[sum_initial]
0.31s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[max]
0.31s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[min_initial]
0.30s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[max_empty]
0.29s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[min]
0.28s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[prod_initial]
0.26s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[mean_empty_int]
0.26s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[mean]
0.26s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[prod]
0.25s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[sum]
0.18s call     rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[mean_empty_float]
0.03s setup    rbc/tests/test_omnisci_array_methods.py::test_ndarray_methods[fill]
```